+++ navycut/__init__.py
@@ -5,7 +5,7 @@ Introduction :
 
 
 """
-version_info = (0, 0, 5)
+version_info = (0, 1, 0)
 
 
 __version__ = ".".join([str(v) for v in version_info])