+++ navycut/boiler_create_app/sister.py
@@ -5,7 +5,7 @@ how to handle it, otherwise it may mess the server.
 
 from navycut.core import AppSister
 from navycut.utils import path
-from .urls import url_patterns
+from .urls import urlpatterns
 
 
 __basedir__ = path.abspath(__file__).parent
@@ -17,5 +17,5 @@ class classname___boiler_var(AppSister):
     static_folder = __basedir__ / "static"
     static_url_path = "/static"
     url_prefix = "/import_name___boiler_var"
-    url_pattern = (url_patterns, )
+    url_pattern = (urlpatterns, )
     import_app_feature = True