+++ navycut/boiler_create_app/urls.py
@@ -1,6 +1,6 @@
 from navycut.urls import path
 # from . import views
 
-url_patterns = [
+urlpatterns = [
     # path("" , views.IndexView, "index"),
 ]