+++ navycut/contrib/admin/urls.py
@@ -5,7 +5,7 @@ import typing as t
 if t.TYPE_CHECKING:
     from navycut.typing import ncUrlPatterns
 
-url_patterns:"ncUrlPatterns" = [
+urlpatterns:"ncUrlPatterns" = [
 
     url("/admin/login/", views.admin_login, "admin_login"),
     url("/admin/logout/", views.admin_logout, "admin_logout")