+++ CHANGELOG.md
@@ -25,6 +25,9 @@
 
 ## [0.1.0] [Upcomming]
 - Several bug fix.
+- Python 3.10.0 compatibility.
+- Removed the `blueprint` support for custom sister apps.
+- Several dependancy update.
 - Removed the support of the relation field like Django.
 - Added the default SQLAlchemy based relation system.
 - Removed the `required` argument from the `Image` field.